+++ doc/WHATSNEW.md
@@ -1,7 +1,10 @@
-Version ??? (unreleased)
-------------------------
+Version 0.12
+------------
 
 - Improved: better sorting of results (eg. mariadb 10 now comes after mariadb 5)
+
+- Improved: dbconsole command: `sql:execute` learned to replace tokens in the value of `--output-file`, eg:
+      `--output-file '../shared/results_{instancename}.yml'`
 
 - New: dbstack learned a new command: `monitor`. It will open an interactive, textual admin console, which gives useful
   information about the running containers (eg. cpu and memory usage)